requirements.txt
Django==5.0
asgiref==3.7.2
django-cors-headers==4.3.0
djangorestframework==3.14.0
# mssql-django==1.3
pyodbc==5.0.1
python-decouple==3.8
pytz==2023.3.post1
setuptools==68.0.0
sqlparse==0.4.4
wheel==0.37.1
